extract configuration logic in its own file
